rearchitect Profile and User as runtime types
